tach file ra cho de quan ly
